backend/requirements.txt
Flask
flask-cors
tensorflow==2.15.0
numpy
Pillow
python-dotenv
requests
gunicorn
opencv-python-headless
